Fixes Issue #4 (ads crash the game)
Ads were crashing the game in a seemingly haphazard way.
After investigation it was found that trying to load an
ad while the current ad is hidden was causing the crash:
A/libc: Fatal signal 11 (SIGSEGV) at 0x00000038 (code=1)

This commit changes the logic so an ad is only attempted
to be loaded while the ad is visible.

Signed-off-by: Isaac Assegai <Isaac.A.Travers@gmail.com>